+++ native/SandSound/App.xaml.cs
@@ -1,0 +1,20 @@
+using Microsoft.UI.Xaml;
+
+namespace SandSound;
+
+public partial class App : Application
+{
+    private Window? _window;
+
+    public App()
+    {
+        InitializeComponent();
+        UnhandledException += (_, args) => Services.AppLog.Write("Unhandled UI exception", args.Exception);
+    }
+
+    protected override void OnLaunched(LaunchActivatedEventArgs args)
+    {
+        _window = new MainWindow();
+        _window.Activate();
+    }
+}